Modified some modules for npm.
